Drop references to slavery
I hope this isn't taken as hypersensitivity or negativity but I know
some people would take umbrage to the default message "Back to work,
slave!", so I dropped the slave part. I realize that no ill will was
meant by this phrase and I get the joke, I just think its better off
removed. I realize this is customizable and did so on my own install,
but it seems better to err on the side of caution here, especially when
the phrase in question doesn't really contribute to the functionality or
usability. Thanks :)